Move dict utils under utils/dict.py
