+++ examples/development/variants.py
@@ -5,7 +5,7 @@ from ray import tune
 import numpy as np
 
 from softlearning.utils.git import get_git_rev
-from softlearning.utils.misc import deep_update
+from softlearning.utils.dict import deep_update
 
 DEFAULT_KEY = "__DEFAULT_KEY__"
