+++ softlearning/utils/misc.py
@@ -1,4 +1,3 @@
-import collections
 import os
 import random
 
@@ -20,46 +19,3 @@ def set_seed(seed):
     np.random.seed(seed)
     tf.compat.v1.set_random_seed(seed)
     print("Using seed {}".format(seed))
-
-
-def deep_update(d, *us):
-    d = d.copy()
-
-    for u in us:
-        u = u.copy()
-        for k, v in u.items():
-            d[k] = (
-                deep_update(d.get(k, {}), v)
-                if isinstance(v, collections.Mapping)
-                else v)
-
-    return d
-
-
-def flatten(unflattened, parent_key='', separator='.'):
-    items = []
-    for k, v in unflattened.items():
-        if separator in k:
-            raise ValueError(
-                "Found separator ({}) from key ({})".format(separator, k))
-        new_key = parent_key + separator + k if parent_key else k
-        if isinstance(v, collections.MutableMapping) and v:
-            items.extend(flatten(v, new_key, separator=separator).items())
-        else:
-            items.append((new_key, v))
-
-    return dict(items)
-
-
-def unflatten(flattened, separator='.'):
-    result = {}
-    for key, value in flattened.items():
-        parts = key.split(separator)
-        d = result
-        for part in parts[:-1]:
-            if part not in d:
-                d[part] = {}
-            d = d[part]
-        d[parts[-1]] = value
-
-    return result
